Works with assets and logs
