+++ modules/core/owner/config/schema/farm_owner.schema.yml
@@ -1,4 +1,4 @@
 # Schema for actions.
-action.configuration.log_assign_action:
+action.configuration.assign_action:
   type: action_configuration_default
-  label: 'Configuration for the log assign action'
+  label: 'Configuration for the assign action'
